feat(insights): platformize summary pages
Replace the HTTP domain, queues destination, and resources summary page
implementations with thin wrappers that render the corresponding prebuilt
dashboards via PrebuiltDashboardRenderer. The landing pages in these
modules were already platformized; this closes the gap for the remaining
detail/summary pages.

Follows the same pattern used for the database span summary page.

Refs DAIN-1568
