+++ static/app/views/insights/browser/resources/views/platformizedSummaryPage.tsx
@@ -1,0 +1,22 @@
+import {DataCategory} from 'sentry/types/core';
+import {useMaxPickableDays} from 'sentry/utils/useMaxPickableDays';
+import {PrebuiltDashboardRenderer} from 'sentry/views/dashboards/prebuiltDashboardRenderer';
+import {PrebuiltDashboardId} from 'sentry/views/dashboards/utils/prebuiltConfigs';
+import {ModulePageProviders} from 'sentry/views/insights/common/components/modulePageProviders';
+
+export function PlatformizedResourceSummaryPage() {
+  const maxPickableDays = useMaxPickableDays({
+    dataCategories: [DataCategory.SPANS],
+  });
+
+  return (
+    <ModulePageProviders
+      moduleName="resource"
+      maxPickableDays={maxPickableDays.maxPickableDays}
+    >
+      <PrebuiltDashboardRenderer
+        prebuiltId={PrebuiltDashboardId.FRONTEND_ASSETS_SUMMARY}
+      />
+    </ModulePageProviders>
+  );
+}
