--- static/app/views/insights/http/views/httpDomainSummaryPage.spec.tsx
@@ -1,399 +1,0 @@
-import {OrganizationFixture} from 'sentry-fixture/organization';
-import {TimeSeriesFixture} from 'sentry-fixture/timeSeries';
-
-import {
-  render,
-  screen,
-  waitFor,
-  waitForElementToBeRemoved,
-} from 'sentry-test/reactTestingLibrary';
-
-import {PageFiltersStore} from 'sentry/components/pageFilters/store';
-import {useReleaseStats} from 'sentry/utils/useReleaseStats';
-import {SAMPLING_MODE} from 'sentry/views/explore/hooks/useProgressiveQuery';
-import {HTTPDomainSummaryPage} from 'sentry/views/insights/http/views/httpDomainSummaryPage';
-
-jest.mock('sentry/utils/useReleaseStats');
-
-describe('HTTPDomainSummaryPage', () => {
-  const organization = OrganizationFixture({features: ['insight-modules']});
-
-  let throughputRequestMock!: jest.Mock;
-  let durationRequestMock!: jest.Mock;
-  let statusRequestMock!: jest.Mock;
-
-  let domainTransactionsListRequestMock: jest.Mock;
-  let domainMetricsRibbonRequestMock: jest.Mock;
-  let regionFilterRequestMock: jest.Mock;
-
-  const initialRouterConfig = {
-    location: {
-      pathname: `/organizations/${organization.slug}/insights/backend/http/domains/`,
-      query: {domain: '*.sentry.dev', statsPeriod: '10d', transactionsCursor: '0:20:0'},
-    },
-    route: '/organizations/:orgId/insights/backend/http/domains/',
-  };
-
-  jest.mocked(useReleaseStats).mockReturnValue({
-    isLoading: false,
-    isPending: false,
-    isError: false,
-    error: null,
-    releases: [],
-  });
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    PageFiltersStore.onInitializeUrlState({
-      projects: [],
-      environments: [],
-      datetime: {
-        period: '10d',
-        start: null,
-        end: null,
-        utc: false,
-      },
-    });
-
-    regionFilterRequestMock = MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/events/',
-      method: 'GET',
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.user-geo-subregion-selector',
-        }),
-      ],
-      body: {
-        data: [
-          {'user.geo.subregion': '21', 'count()': 123},
-          {'user.geo.subregion': '155', 'count()': 123},
-        ],
-        meta: {
-          fields: {'user.geo.subregion': 'string', 'count()': 'integer'},
-        },
-      },
-    });
-
-    domainTransactionsListRequestMock = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events/`,
-      method: 'GET',
-      body: {
-        data: [],
-      },
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-transactions-list',
-        }),
-      ],
-    });
-
-    domainMetricsRibbonRequestMock = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events/`,
-      method: 'GET',
-      body: {
-        data: [],
-      },
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-metrics-ribbon',
-        }),
-      ],
-    });
-
-    throughputRequestMock = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-timeseries/`,
-      method: 'GET',
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-throughput-chart',
-        }),
-      ],
-      body: {
-        timeSeries: [
-          TimeSeriesFixture({
-            yAxis: 'epm()',
-            values: [
-              {value: 7810.2, timestamp: 1699907700000},
-              {value: 1216.8, timestamp: 1699908000000},
-            ],
-          }),
-        ],
-      },
-    });
-
-    durationRequestMock = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-timeseries/`,
-      method: 'GET',
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-duration-chart',
-        }),
-      ],
-      body: {
-        timeSeries: [
-          TimeSeriesFixture({
-            yAxis: 'avg(span.self_time)',
-            values: [
-              {value: 710.2, timestamp: 1699907700000},
-              {value: 116.8, timestamp: 1699908000000},
-            ],
-          }),
-        ],
-      },
-    });
-
-    statusRequestMock = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-timeseries/`,
-      method: 'GET',
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-response-code-chart',
-        }),
-      ],
-      body: {
-        timeSeries: [
-          TimeSeriesFixture({
-            yAxis: 'http_response_rate(3)',
-            values: [{value: 0.2, timestamp: 1699908000000}],
-          }),
-          TimeSeriesFixture({
-            yAxis: 'http_response_rate(4)',
-            values: [{value: 0.1, timestamp: 1699908000000}],
-          }),
-          TimeSeriesFixture({
-            yAxis: 'http_response_rate(5)',
-            values: [{value: 0.3, timestamp: 1699908000000}],
-          }),
-        ],
-      },
-    });
-  });
-
-  afterAll(() => {
-    jest.resetAllMocks();
-  });
-
-  it('fetches module data', async () => {
-    render(<HTTPDomainSummaryPage />, {organization, initialRouterConfig});
-
-    await waitFor(() => {
-      expect(throughputRequestMock).toHaveBeenNthCalledWith(
-        1,
-        `/organizations/${organization.slug}/events-timeseries/`,
-        expect.objectContaining({
-          method: 'GET',
-          query: {
-            dataset: 'spans',
-            sampling: SAMPLING_MODE.NORMAL,
-            environment: [],
-            excludeOther: 0,
-            groupBy: undefined,
-            interval: '30m',
-            partial: 1,
-            project: [],
-            query: 'span.op:http.client span.domain:"\\*.sentry.dev"',
-            referrer: 'api.insights.http.domain-summary-throughput-chart',
-            statsPeriod: '10d',
-            yAxis: ['epm()'],
-            caseInsensitive: undefined,
-          },
-        })
-      );
-    });
-
-    expect(durationRequestMock).toHaveBeenNthCalledWith(
-      1,
-      `/organizations/${organization.slug}/events-timeseries/`,
-      expect.objectContaining({
-        method: 'GET',
-        query: {
-          dataset: 'spans',
-          sampling: SAMPLING_MODE.NORMAL,
-          environment: [],
-          excludeOther: 0,
-          groupBy: undefined,
-          interval: '30m',
-          partial: 1,
-          project: [],
-          query: 'span.op:http.client span.domain:"\\*.sentry.dev"',
-          referrer: 'api.insights.http.domain-summary-duration-chart',
-          statsPeriod: '10d',
-          yAxis: ['avg(span.self_time)'],
-          caseInsensitive: undefined,
-        },
-      })
-    );
-
-    expect(statusRequestMock).toHaveBeenNthCalledWith(
-      1,
-      `/organizations/${organization.slug}/events-timeseries/`,
-      expect.objectContaining({
-        method: 'GET',
-        query: {
-          dataset: 'spans',
-          sampling: SAMPLING_MODE.NORMAL,
-          environment: [],
-          excludeOther: 0,
-          groupBy: undefined,
-          interval: '30m',
-          partial: 1,
-          project: [],
-          query: 'span.op:http.client span.domain:"\\*.sentry.dev"',
-          referrer: 'api.insights.http.domain-summary-response-code-chart',
-          statsPeriod: '10d',
-          yAxis: [
-            'http_response_rate(3)',
-            'http_response_rate(4)',
-            'http_response_rate(5)',
-          ],
-          caseInsensitive: undefined,
-        },
-      })
-    );
-
-    expect(domainMetricsRibbonRequestMock).toHaveBeenNthCalledWith(
-      1,
-      `/organizations/${organization.slug}/events/`,
-      expect.objectContaining({
-        method: 'GET',
-        query: {
-          dataset: 'spans',
-          environment: [],
-          field: [
-            'epm()',
-            'avg(span.self_time)',
-            'sum(span.self_time)',
-            'http_response_rate(3)',
-            'http_response_rate(4)',
-            'http_response_rate(5)',
-          ],
-          per_page: 50,
-          project: [],
-          query: 'span.op:http.client span.domain:"\\*.sentry.dev"',
-          referrer: 'api.insights.http.domain-summary-metrics-ribbon',
-          sampling: SAMPLING_MODE.NORMAL,
-          statsPeriod: '10d',
-        },
-      })
-    );
-
-    expect(domainTransactionsListRequestMock).toHaveBeenNthCalledWith(
-      1,
-      `/organizations/${organization.slug}/events/`,
-      expect.objectContaining({
-        method: 'GET',
-        query: {
-          dataset: 'spans',
-          environment: [],
-          field: [
-            'project.id',
-            'transaction',
-            'transaction.method',
-            'epm()',
-            'http_response_rate(3)',
-            'http_response_rate(4)',
-            'http_response_rate(5)',
-            'avg(span.self_time)',
-            'sum(span.self_time)',
-          ],
-          per_page: 20,
-          project: [],
-          cursor: '0:20:0',
-          query: 'span.op:http.client span.domain:"\\*.sentry.dev"',
-          sort: '-sum(span.self_time)',
-          referrer: 'api.insights.http.domain-summary-transactions-list',
-          sampling: SAMPLING_MODE.NORMAL,
-          statsPeriod: '10d',
-        },
-      })
-    );
-
-    expect(regionFilterRequestMock).toHaveBeenCalledWith(
-      `/organizations/${organization.slug}/events/`,
-      expect.objectContaining({
-        method: 'GET',
-        query: {
-          dataset: 'spans',
-          environment: [],
-          field: ['user.geo.subregion', 'count()'],
-          per_page: 50,
-          project: [],
-          query: 'has:user.geo.subregion',
-          sort: '-count()',
-          referrer: 'api.insights.user-geo-subregion-selector',
-          sampling: SAMPLING_MODE.NORMAL,
-          statsPeriod: '10d',
-        },
-      })
-    );
-
-    await waitForElementToBeRemoved(() => screen.queryAllByTestId('loading-indicator'));
-  });
-
-  it('renders a list of queries', async () => {
-    MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events/`,
-      method: 'GET',
-
-      match: [
-        MockApiClient.matchQuery({
-          referrer: 'api.insights.http.domain-summary-transactions-list',
-        }),
-      ],
-      body: {
-        data: [
-          {
-            'project.id': 8,
-            transaction: '/api/users',
-            'transaction.method': 'GET',
-            'epm()': 17.88,
-            'http_response_rate(3)': 0.97,
-            'http_response_rate(4)': 0.025,
-            'http_response_rate(5)': 0.005,
-            'avg(span.self_time)': 204.5,
-            'sum(span.self_time)': 177238,
-          },
-        ],
-        meta: {
-          fields: {
-            'epm()': 'rate',
-            'avg(span.self_time)': 'duration',
-            'http_response_rate(3)': 'percentage',
-            'http_response_rate(4)': 'percentage',
-            'http_response_rate(5)': 'percentage',
-            'sum(span.self_time)': 'duration',
-          },
-        },
-      },
-    });
-
-    render(<HTTPDomainSummaryPage />, {organization, initialRouterConfig});
-
-    await waitForElementToBeRemoved(() => screen.queryAllByTestId('loading-indicator'));
-
-    expect(screen.getByRole('table', {name: 'Transactions'})).toBeInTheDocument();
-
-    expect(screen.getByRole('columnheader', {name: 'Found In'})).toBeInTheDocument();
-    expect(
-      screen.getByRole('columnheader', {name: 'Requests Per Minute'})
-    ).toBeInTheDocument();
-    expect(screen.getByRole('columnheader', {name: '3XXs'})).toBeInTheDocument();
-    expect(screen.getByRole('columnheader', {name: '4XXs'})).toBeInTheDocument();
-    expect(screen.getByRole('columnheader', {name: '5XXs'})).toBeInTheDocument();
-    expect(screen.getByRole('columnheader', {name: 'Avg Duration'})).toBeInTheDocument();
-    expect(screen.getByRole('columnheader', {name: 'Time Spent'})).toBeInTheDocument();
-
-    expect(screen.getByRole('cell', {name: 'GET /api/users'})).toBeInTheDocument();
-    expect(screen.getByRole('link', {name: 'GET /api/users'})).toHaveAttribute(
-      'href',
-      '/organizations/org-slug/insights/backend/http/domains/?domain=%2A.sentry.dev&project=8&statsPeriod=10d&transaction=%2Fapi%2Fusers&transactionMethod=GET&transactionsCursor=0%3A20%3A0'
-    );
-    expect(screen.getByRole('cell', {name: '17.9/s'})).toBeInTheDocument();
-    expect(screen.getByRole('cell', {name: '97%'})).toBeInTheDocument();
-    expect(screen.getByRole('cell', {name: '2.5%'})).toBeInTheDocument();
-    expect(screen.getByRole('cell', {name: '0.5%'})).toBeInTheDocument();
-    expect(screen.getByRole('cell', {name: '204.50ms'})).toBeInTheDocument();
-    expect(screen.getByRole('cell', {name: '2.95min'})).toBeInTheDocument();
-  });
-});
